- Improvement component schema of revisionDiff feature
- Refactoring reviewDiffContainer and reviewDiff components

- Refactoring ReviewSidePanel for paypal
