+++ public/js/cat_source/es6/react/components/review_extended/ReviewExtendedPanel.js
@@ -2,6 +2,7 @@ import ReviewVersionsDiff from "./ReviewVersionsDiff";
 
 let ReviewExtendedIssuesContainer = require('./ReviewExtendedIssuesContainer').default;
 let ReviewVersionDiffContainer = require('./ReviewVersionsDiffContainer').default;
+let ReviewVersionDiff = require('./ReviewVersionsDiff').default;
 let ReviewExtendedIssuePanel = require('./ReviewExtendedIssuePanel').default;
 let SegmentConstants = require('../../constants/SegmentConstants');
 let SegmentStore = require('../../stores/SegmentStore');
@@ -17,21 +18,12 @@ class ReviewExtendedPanel extends React.Component {
 
 	}
 
-	trackChanges(sid, editareaText) {
-		let text = htmlEncode(UI.prepareTextToSend(editareaText));
-		if (this.props.segment.sid === sid) {
-			this.setState({
-				translation: text,
-			});
-		}
-	}
-
 	textSelected(data, diffPatch) {
-		/*this.setState({
+		this.setState({
 			addIssue: true,
 			selectionObj: data,
 			diffPatch: diffPatch
-		});*/
+		});
 	}
 
 	removeSelection() {
@@ -61,73 +53,19 @@ class ReviewExtendedPanel extends React.Component {
 	}
 
 	componentWillReceiveProps(nextProps) {
-		this.originalTranslation = htmlEncode(UI.prepareTextToSend(nextProps.segment.translation));
 		this.setState({
-			translation: this.originalTranslation,
 			addIssue: false,
-			selectionObj: null,
-			diffPatch: null,
 			versionNumber: nextProps.segment.versions[0].version_number
 		});
 	}
 
-	componentDidMount() {
-		SegmentStore.addListener(SegmentConstants.TRANSLATION_EDITED, this.trackChanges.bind(this));
-	}
-
-	componentWillUnmount() {
-		SegmentStore.removeListener(SegmentConstants.TRANSLATION_EDITED, this.trackChanges);
-	}
-
 	render() {
 		let issues = this.getAllIssues();
-		/*return <div className="review-issues-overview-panel">
-			 <div className="review-version-wrapper">
-				 <div className="review-translation-version" >
-					 <div className="review-version-header">
-						 <h3>Live changes</h3>
-					 </div>
-					 <div className="collapsable">
-
-						 <ReviewVersionDiff
-							 textSelectedFn={this.textSelected.bind(this)}
-							 removeSelection={this.removeSelection.bind(this)}
-							 previousVersion={this.originalTranslation}
-							 translation={this.state.translation}
-							 segment={this.props.segment}
-							 decodeTextFn={UI.decodeText}
-							 selectable={this.props.isReview}
-						 />
-
-
-						 <div className="error-type">
-							 <ReviewExtendedIssuePanel
-								 sid={this.props.segment.sid}
-								 selection={this.state.selectionObj}
-								 segmentVersion={this.state.versionNumber}
-								 diffPatch={this.state.diffPatch}
-								 submitIssueCallback={this.removeSelection.bind(this)}
-								 reviewType={this.props.reviewType}
-								 segment={this.props.segment}
-							 />
-						 </div>
-						 <ReviewExtendedIssuesContainer
-							 issueMouseEnter={this.issueMouseEnter.bind(this)}
-							 issueMouseLeave={this.issueMouseLeave.bind(this)}
-							 reviewType={this.props.reviewType}
-							 issues={issues}
-							 sid={this.props.segment.sid}
-						 />
-					 </div>
-				 </div>
-			 </div>
-		 </div>;*/
 
 		return <div className="re-wrapper">
 			<ReviewVersionDiffContainer
 				textSelectedFn={this.textSelected.bind(this)}
 				removeSelection={this.removeSelection.bind(this)}
-				translation={this.state.translation}
 				segment={this.props.segment}
 				selectable={this.props.isReview}
 			/>
@@ -140,6 +78,7 @@ class ReviewExtendedPanel extends React.Component {
 				reviewType={this.props.reviewType}
 				segment={this.props.segment}
 			/>
+
 			<ReviewExtendedIssuesContainer
 				issueMouseEnter={this.issueMouseEnter.bind(this)}
 				issueMouseLeave={this.issueMouseLeave.bind(this)}
