+++ public/js/cat_source/es6/react/components/review_extended/ReviewVersionsDiffContainer.js
@@ -4,6 +4,7 @@
  */
 let ReviewVersionDiff =  require("./ReviewVersionsDiff").default;
 let React = require('react');
+let SegmentConstants = require('../../constants/SegmentConstants');
 class ReviewVersionsDiffContainer extends React.Component {
 
     constructor(props) {
@@ -13,19 +14,38 @@ class ReviewVersionsDiffContainer extends React.Component {
 			translation: this.props.segment.translation,
 			selectionObj: null,
 			diffPatch: null,
+			sid: this.props.segment.sid
 		};
     }
 
+	issueMouseEnter(issue, event, reactid) {
+		SegmentActions.showSelection(this.props.sid, issue);
+	}
+
+	trackChanges(sid, editareaText) {
+		let text = htmlEncode(UI.prepareTextToSend(editareaText));
+		if (this.props.segment.sid === sid) {
+			this.setState({
+				translation: text,
+			});
+		}
+	}
+
+	componentWillReceiveProps(nextProps) {
+		this.originalTranslation = htmlEncode(UI.prepareTextToSend(nextProps.segment.translation));
+		this.setState({
+			translation: this.originalTranslation,
+			selectionObj: null,
+			diffPatch: null
+		});
+	}
 
     componentDidMount() {
+		SegmentStore.addListener(SegmentConstants.TRANSLATION_EDITED, this.trackChanges.bind(this));
     }
 
     componentWillUnmount() {
-    }
-    componentWillMount() {}
-
-    componentDidUpdate() {
-
+		SegmentStore.removeListener(SegmentConstants.TRANSLATION_EDITED, this.trackChanges);
     }
 
     allowHTML(string) {
@@ -50,7 +70,7 @@ class ReviewVersionsDiffContainer extends React.Component {
 						translation={this.state.translation}
 						segment={this.props.segment}
 						decodeTextFn={UI.decodeText}
-						selectable={this.props.isReview}
+						selectable={this.props.selectable}
 					/>
 				</div>
 			</div>
